@@ -14,7 +14,7 @@ const simplePeriodicTask = "simplePeriodicTask";
 const simplePeriodic1HourTask = "simplePeriodic1HourTask";
 
 void callbackDispatcher() {
-  Workmanager.executeTask((task, inputData) async {
+  Workmanager().executeTask((task, inputData) async {
     switch (task) {
       case simpleTaskKey:
         print("$simpleTaskKey was executed. inputData = $inputData");
@@ -33,8 +33,8 @@ void callbackDispatcher() {
         break;
       case Workmanager.iOSBackgroundTask:
         print("The iOS background fetch was triggered");
-        Directory tempDir = await getTemporaryDirectory();
-        String tempPath = tempDir.path;
+        Directory? tempDir = await getTemporaryDirectory();
+        String? tempPath = tempDir?.path;
         print(
             "You can access other plugins in the background, for example Directory.getTemporaryDirectory(): $tempPath");
         break;
@@ -55,16 +55,15 @@ class PlatformEnabledButton extends RaisedButton {
   final _Platform platform;
 
   PlatformEnabledButton({
-    this.platform,
-    @required Widget child,
-    @required VoidCallback onPressed,
-  })  : assert(child != null, onPressed != null),
-        super(
+    required this.platform,
+    required Widget child,
+    required VoidCallback onPressed,
+  }) : super(
             child: child,
-            onPressed:
-                (Platform.isAndroid && platform == _Platform.android || Platform.isIOS && platform == _Platform.ios)
-                    ? onPressed
-                    : null);
+            onPressed: (Platform.isAndroid && platform == _Platform.android ||
+                    Platform.isIOS && platform == _Platform.ios)
+                ? onPressed
+                : null);
 }
 
 class _MyAppState extends State<MyApp> {
@@ -80,24 +79,26 @@ class _MyAppState extends State<MyApp> {
           child: Column(
             crossAxisAlignment: CrossAxisAlignment.stretch,
             children: <Widget>[
-              Text("Plugin initialization", style: Theme.of(context).textTheme.headline),
+              Text("Plugin initialization",
+                  style: Theme.of(context).textTheme.headline),
               RaisedButton(
                   child: Text("Start the Flutter background service"),
                   onPressed: () {
-                    Workmanager.initialize(
+                    Workmanager().initialize(
                       callbackDispatcher,
                       isInDebugMode: true,
                     );
                   }),
               SizedBox(height: 16),
-              Text("One Off Tasks (Android only)", style: Theme.of(context).textTheme.headline),
+              Text("One Off Tasks (Android only)",
+                  style: Theme.of(context).textTheme.headline),
               //This task runs once.
               //Most likely this will trigger immediately
               PlatformEnabledButton(
                   platform: _Platform.android,
                   child: Text("Register OneOff Task"),
                   onPressed: () {
-                    Workmanager.registerOneOffTask(
+                    Workmanager().registerOneOffTask(
                       "1",
                       simpleTaskKey,
                       inputData: <String, dynamic>{
@@ -115,14 +116,15 @@ class _MyAppState extends State<MyApp> {
                   platform: _Platform.android,
                   child: Text("Register Delayed OneOff Task"),
                   onPressed: () {
-                    Workmanager.registerOneOffTask(
+                    Workmanager().registerOneOffTask(
                       "2",
                       simpleDelayedTask,
                       initialDelay: Duration(seconds: 10),
                     );
                   }),
               SizedBox(height: 8),
-              Text("Periodic Tasks (Android only)", style: Theme.of(context).textTheme.headline),
+              Text("Periodic Tasks (Android only)",
+                  style: Theme.of(context).textTheme.headline),
               //This task runs periodically
               //It will wait at least 10 seconds before its first launch
               //Since we have not provided a frequency it will be the default 15 minutes
@@ -130,7 +132,7 @@ class _MyAppState extends State<MyApp> {
                   platform: _Platform.android,
                   child: Text("Register Periodic Task"),
                   onPressed: () {
-                    Workmanager.registerPeriodicTask(
+                    Workmanager().registerPeriodicTask(
                       "3",
                       simplePeriodicTask,
                       initialDelay: Duration(seconds: 10),
@@ -142,7 +144,7 @@ class _MyAppState extends State<MyApp> {
                   platform: _Platform.android,
                   child: Text("Register 1 hour Periodic Task"),
                   onPressed: () {
-                    Workmanager.registerPeriodicTask(
+                    Workmanager().registerPeriodicTask(
                       "5",
                       simplePeriodic1HourTask,
                       frequency: Duration(hours: 1),
@@ -152,7 +154,7 @@ class _MyAppState extends State<MyApp> {
                 platform: _Platform.android,
                 child: Text("Cancel All"),
                 onPressed: () async {
-                  await Workmanager.cancelAll();
+                  await Workmanager().cancelAll();
                   print('Cancel all tasks completed');
                 },
               ),
