@@ -16,14 +16,15 @@ private interface CallHandler<T : WorkManagerCall> {
     fun handle(context: Context, convertedCall: T, result: MethodChannel.Result)
 }
 
-class WorkmanagerCallHandler(private val ctx: Context) {
-    fun handle(call: MethodCall, result: MethodChannel.Result) =
-            when (val extractedCall = Extractor.extractWorkManagerCallFromRawMethodName(call)) {
-                is WorkManagerCall.Initialize -> InitializeHandler.handle(ctx, extractedCall, result)
-                is WorkManagerCall.RegisterTask -> RegisterTaskHandler.handle(ctx, extractedCall, result)
-                is WorkManagerCall.CancelTask -> UnregisterTaskHandler.handle(ctx, extractedCall, result)
-                is WorkManagerCall.Unknown -> UnknownTaskHandler.handle(ctx, extractedCall, result)
-            }
+class WorkmanagerCallHandler(private val ctx: Context) : MethodChannel.MethodCallHandler {
+    override fun onMethodCall(call: MethodCall, result: MethodChannel.Result) {
+        when (val extractedCall = Extractor.extractWorkManagerCallFromRawMethodName(call)) {
+            is WorkManagerCall.Initialize -> InitializeHandler.handle(ctx, extractedCall, result)
+            is WorkManagerCall.RegisterTask -> RegisterTaskHandler.handle(ctx, extractedCall, result)
+            is WorkManagerCall.CancelTask -> UnregisterTaskHandler.handle(ctx, extractedCall, result)
+            is WorkManagerCall.Unknown -> UnknownTaskHandler.handle(ctx, extractedCall, result)
+        }
+    }
 }
 
 private object InitializeHandler : CallHandler<WorkManagerCall.Initialize> {
