@@ -15,7 +15,7 @@ const _noDuration = const Duration(seconds: 0);
 /// [taskName] Returns the value you provided when registering the task.
 /// iOS will always return [Workmanager.iOSBackgroundTask]
 typedef BackgroundTaskHandler = Future<bool> Function(
-    String taskName, Map<String, dynamic> inputData);
+    String taskName, Map<String, dynamic>? inputData);
 
 /// Make sure you followed the platform setup steps first before trying to register any task.
 /// Android:
@@ -49,6 +49,19 @@ typedef BackgroundTaskHandler = Future<bool> Function(
 ///
 /// iOS periodic task is automatically scheduled if you setup the plugin properly.
 class Workmanager {
+  factory Workmanager() => _instance;
+
+  Workmanager._internal(
+      MethodChannel backgroundChannel, MethodChannel foregroundChannel)
+      : _backgroundChannel = backgroundChannel,
+        _foregroundChannel = foregroundChannel;
+
+  static final Workmanager _instance = Workmanager._internal(
+      const MethodChannel(
+          "be.tramckrijte.workmanager/background_channel_work_manager"),
+      const MethodChannel(
+          "be.tramckrijte.workmanager/foreground_channel_work_manager"));
+
   /// Use this constant inside your callbackDispatcher to identify when an iOS Background Fetch occurred.
   ///
   /// ```
@@ -67,13 +80,13 @@ class Workmanager {
   static const String iOSBackgroundTask = "iOSPerformFetch";
   static bool _isInDebugMode = false;
 
-  static const MethodChannel _backgroundChannel = const MethodChannel(
+  MethodChannel _backgroundChannel = const MethodChannel(
       "be.tramckrijte.workmanager/background_channel_work_manager");
-  static const MethodChannel _foregroundChannel = const MethodChannel(
+  MethodChannel _foregroundChannel = const MethodChannel(
       "be.tramckrijte.workmanager/foreground_channel_work_manager");
 
   /// A helper function so you only need to implement a [BackgroundTaskHandler]
-  static void executeTask(final BackgroundTaskHandler backgroundTask) {
+  void executeTask(final BackgroundTaskHandler backgroundTask) {
     WidgetsFlutterBinding.ensureInitialized();
     _backgroundChannel.setMethodCallHandler((call) async {
       final inputData = call.arguments["be.tramckrijte.workmanager.INPUT_DATA"];
@@ -88,7 +101,7 @@ class Workmanager {
   /// This call is required if you wish to use the [WorkManager] plugin.
   /// [callbackDispatcher] is a top level function which will be invoked by Android
   /// [isInDebugMode] true will post debug notifications with information about when a task should have run
-  static Future<void> initialize(
+  Future<void> initialize(
     final Function callbackDispatcher, {
     final bool isInDebugMode = false,
   }) async {
@@ -96,30 +109,32 @@ class Workmanager {
     final callback = PluginUtilities.getCallbackHandle(callbackDispatcher);
     assert(callback != null,
         "The callbackDispatcher needs to be either a static function or a top level function to be accessible as a Flutter entry point.");
-    final int handle = callback.toRawHandle();
-    await _foregroundChannel.invokeMethod(
-      'initialize',
-      JsonMapperHelper.toInitializeMethodArgument(
-        isInDebugMode: _isInDebugMode,
-        callbackHandle: handle,
-      ),
-    );
+    if (callback != null) {
+      final int handle = callback.toRawHandle();
+      await _foregroundChannel.invokeMethod<void>(
+        'initialize',
+        JsonMapperHelper.toInitializeMethodArgument(
+          isInDebugMode: _isInDebugMode,
+          callbackHandle: handle,
+        ),
+      );
+    }
   }
 
   /// Schedule a one off task
   /// A [uniqueName] is required so only one task can be registered.
   /// The [taskName] is the value that will be returned in the [BackgroundTaskHandler]
   /// The [inputData] is the input data for task. Valid value types are: int, bool, double, String and their list
-  static Future<void> registerOneOffTask(
+  Future<void> registerOneOffTask(
     final String uniqueName,
     final String taskName, {
-    final String tag,
-    final ExistingWorkPolicy existingWorkPolicy,
+    final String? tag,
+    final ExistingWorkPolicy? existingWorkPolicy,
     final Duration initialDelay = _noDuration,
-    final Constraints constraints,
-    final BackoffPolicy backoffPolicy,
+    final Constraints? constraints,
+    final BackoffPolicy? backoffPolicy,
     final Duration backoffPolicyDelay = _noDuration,
-    final Map<String, dynamic> inputData,
+    final Map<String, dynamic>? inputData,
   }) async =>
       await _foregroundChannel.invokeMethod(
         "registerOneOffTask",
@@ -143,17 +158,17 @@ class Workmanager {
   /// a [frequency] is not required and will be defaulted to 15 minutes if not provided.
   /// a [frequency] has a minimum of 15 min. Android will automatically change your frequency to 15 min if you have configured a lower frequency.
   /// The [inputData] is the input data for task. Valid value types are: int, bool, double, String and their list
-  static Future<void> registerPeriodicTask(
+  Future<void> registerPeriodicTask(
     final String uniqueName,
     final String taskName, {
-    final Duration frequency,
-    final String tag,
-    final ExistingWorkPolicy existingWorkPolicy,
+    final Duration? frequency,
+    final String? tag,
+    final ExistingWorkPolicy? existingWorkPolicy,
     final Duration initialDelay = _noDuration,
-    final Constraints constraints,
-    final BackoffPolicy backoffPolicy,
+    final Constraints? constraints,
+    final BackoffPolicy? backoffPolicy,
     final Duration backoffPolicyDelay = _noDuration,
-    final Map<String, dynamic> inputData,
+    final Map<String, dynamic>? inputData,
   }) async =>
       await _foregroundChannel.invokeMethod(
         "registerPeriodicTask",
@@ -173,39 +188,39 @@ class Workmanager {
       );
 
   /// Cancels a task by its [uniqueName]
-  static Future<void> cancelByUniqueName(final String uniqueName) async =>
+  Future<void> cancelByUniqueName(final String uniqueName) async =>
       await _foregroundChannel.invokeMethod(
         "cancelTaskByUniqueName",
         {"uniqueName": uniqueName},
       );
 
   /// Cancels a task by its [tag]
-  static Future<void> cancelByTag(final String tag) async =>
+  Future<void> cancelByTag(final String tag) async =>
       await _foregroundChannel.invokeMethod(
         "cancelTaskByTag",
         {"tag": tag},
       );
 
   /// Cancels all tasks
-  static Future<void> cancelAll() async =>
+  Future<void> cancelAll() async =>
       await _foregroundChannel.invokeMethod("cancelAllTasks");
 }
 
 /// A helper object to convert the selected options to JSON format. Mainly for testability.
 class JsonMapperHelper {
   @visibleForTesting
-  static Map<String, Object> toRegisterMethodArgument({
-    final bool isInDebugMode,
-    final String uniqueName,
-    final String taskName,
-    final Duration frequency,
-    final String tag,
-    final ExistingWorkPolicy existingWorkPolicy,
-    final Duration initialDelay,
-    final Constraints constraints,
-    final BackoffPolicy backoffPolicy,
-    final Duration backoffPolicyDelay,
-    final Map<String, dynamic> inputData,
+  static Map<String, Object?> toRegisterMethodArgument({
+    final bool isInDebugMode = false,
+    final String? uniqueName,
+    final String? taskName,
+    final Duration? frequency,
+    final String? tag,
+    final ExistingWorkPolicy? existingWorkPolicy,
+    final Duration? initialDelay,
+    final Constraints? constraints,
+    final BackoffPolicy? backoffPolicy,
+    final Duration? backoffPolicyDelay,
+    final Map<String, dynamic>? inputData,
   }) {
     if (inputData != null) {
       for (final entry in inputData.entries) {
@@ -234,32 +249,29 @@ class JsonMapperHelper {
       "tag": tag,
       "frequency": frequency?.inSeconds,
       "existingWorkPolicy": _enumToString(existingWorkPolicy),
-      "initialDelaySeconds": initialDelay.inSeconds,
+      "initialDelaySeconds": initialDelay?.inSeconds,
       "networkType": _enumToString(constraints?.networkType),
       "requiresBatteryNotLow": constraints?.requiresBatteryNotLow,
       "requiresCharging": constraints?.requiresCharging,
       "requiresDeviceIdle": constraints?.requiresDeviceIdle,
       "requiresStorageNotLow": constraints?.requiresStorageNotLow,
       "backoffPolicyType": _enumToString(backoffPolicy),
-      "backoffDelayInMilliseconds": backoffPolicyDelay.inMilliseconds,
-      "inputData": inputData == null
-          ? null
-          : inputData == null ? null : jsonEncode(inputData),
+      "backoffDelayInMilliseconds": backoffPolicyDelay?.inMilliseconds,
+      "inputData": inputData == null ? null : jsonEncode(inputData),
     };
   }
 
   @visibleForTesting
-  static Map<String, Object> toInitializeMethodArgument({
-    final bool isInDebugMode,
-    final int callbackHandle,
+  static Map<String, Object?> toInitializeMethodArgument({
+    required final bool isInDebugMode,
+    required final int callbackHandle,
   }) {
-    assert(callbackHandle != null);
     return {
       "isInDebugMode": isInDebugMode,
       "callbackHandle": callbackHandle,
     };
   }
 
-  static String _enumToString(final dynamic enumeration) =>
+  static String? _enumToString(final dynamic enumeration) =>
       enumeration?.toString()?.split('.')?.last;
 }
